requirements.txt
aiogram==2.25.1
selenium==4.19.0
python-dotenv==1.0.1